ReBuild: Doors and Corners glass patch
Optional patch to enable ReBuild: Doors and Corners glass using sand instead of stone chunks. Tested with only relevant mods installed + all DLC and it seemed to work as fine as all the other patches do (the patching method seems to be a bit janky, but I'm not going to overhaul it for one patch and it works).